move logic from SplashViewModel into Use Case
